EY-2594: Bruker felles konfigurasjon
