+++ apps/etterlatte-egne-ansatte-lytter/src/main/kotlin/kafka/KafkaEnvironment.kt
@@ -1,0 +1,8 @@
+package no.nav.etterlatte.kafka
+
+import org.apache.kafka.common.serialization.StringDeserializer
+
+class KafkaEnvironment : Kafkakonfigurasjon(
+    groupId = "SKJERMING_GROUP_ID",
+    deserializerClass = StringDeserializer::class
+)
